# edge/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Police Incident Management System API
  description: |
    REST API for the Police Incident Management System. This is an event-driven system where 
    all operations produce events to Kafka. Events represent requests/commands from the edge, 
    not state changes.
    
    ## Architecture
    - Edge servers receive HTTP requests and produce events to Kafka
    - Events represent requests/commands, not state changes
    - No state reconstruction in edge layer
    - All operations are asynchronous via Kafka events
    
  version: 1.0.0
  contact:
    name: Police System API Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.policesystem.example.com/api/v1
    description: Production server
...
paths:
...
  /webhooks:
    post:
      tags:
        - Webhooks
      summary: Register webhook subscription
      description: Registers a webhook URL to receive notifications for specified event types
      operationId: registerWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRegistration'
      responses:
        '201':
          description: Webhook subscription created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                          url:
                            type: string
                          events:
                            type: array
                            items:
                              type: string
        '400':
          $ref: '#/components/responses/BadRequest'
    
    get:
      tags:
        - Webhooks
      summary: List webhook subscriptions
      description: Returns all registered webhook subscriptions
      operationId: listWebhooks
      responses:
        '200':
          description: Webhook subscriptions retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
  /webhooks/{id}:
    delete:
      tags:
        - Webhooks
      summary: Delete webhook subscription
      description: Deletes a webhook subscription by ID
      operationId: deleteWebhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook subscription deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # Query
  # ============================================
...
components:
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Request processed successfully"
        data:
          description: Response data (type varies by endpoint)
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
          nullable: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: string
          example: ["badgeNumber is required"]
        validValues:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Map of field names to their valid enum values (only present for enum validation errors)
          example:
            unitType: ["Single", "Team", "Squad"]
    ResponseMetadata:
      type: object
      properties:
        correlationId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
          description: "Correlation ID for request tracing"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:40:00.000+00:00"
        version:
          type: string
          example: "1.0"
        links:
          type: object
          additionalProperties:
            type: string
          description: "HATEOAS-style links to related resources"
          example:
            self: "/api/v1/incidents/INC-001"
            dispatch: "/api/v1/incidents/INC-001/dispatch"
...
    WebhookRegistration:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          format: uri
          example: "https://example.com/webhooks/incidents"
        events:
          type: array
          items:
            type: string
          example: ["ReportIncidentRequested", "DispatchIncidentRequested"]
        secret:
          type: string
          example: "webhook-secret-key"
          description: "Secret for webhook signature verification"
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'